requirements.txt
alembic==1.14.0
aniso8601==10.0.0
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
caldav==1.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
docx2pdf==0.1.8
fake-useragent==2.1.0
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fpdf==1.7.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
greenlet==1.1.2
httplib2==0.22.0
icalendar==6.1.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==5.3.1
Mako==1.3.9
MarkupSafe==2.0.1
marshmallow==3.23.2
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.19.2
pytz==2025.1
pywin32==310
recurring-ical-events==3.6.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.6
SQLAlchemy==1.4.28
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vobject==0.9.9
Werkzeug==2.0.2
WTForms==3.0.0
x-wr-timezone==2.0.1
zipp==3.21.0
